jobs/trigger-agent/requirements.txt
functions-framework==3.*
requests
google-auth
